Redo MFC r300220,r300223:
Differential Revision: https://reviews.freebsd.org/D6803
Reviewed by: alc, kib
Sponsored by: EMC / Isilon Storage Division

r300220 (by cem):

sys/vmmeter.h: Fix trivial '-Wsign-compare' warning in common header

Frankly, it doesn't make sense for vm_pageout_wakeup_thresh to have a negative
value (it is only ever set to a fraction of v_free_min, which is unsigned and
also obviously non-negative). But I'm not going to try and convert every
non-negative scalar in the VM to unsigned today, so just cast it for the
comparison.

r300223 (by cem):

vm/vm_page.h: Fix trivial '-Wpointer-sign' warning

pq_vcnt, as a count of real things, has no business being negative. It is only
ever initialized by a u_int counter.

The warning came from the atomic_add_int() in vm_pagequeue_cnt_add().

Rectify the warning by changing the variable to u_int. No functional change.

Suggested by: Clang 3.3

+++ sys/sys/vmmeter.h
@@ -183,7 +183,8 @@ static __inline
 int
 vm_paging_needed(void)
 {
-    return (cnt.v_free_count + cnt.v_cache_count < vm_pageout_wakeup_thresh);
+    return (cnt.v_free_count + cnt.v_cache_count <
+        (u_int)vm_pageout_wakeup_thresh);
 }
 
 #endif
